not none
Improving your skills
Open source projects come with a lot of demands, so many skills and applications are required to accomplish the tasks. This characteristic accords you as a contributor the opportunity to practice and build on skills you already have, and to acquire new skills as well.

Contributing to open source will allow you to practice in many fields that include (but are not limited to) graphic design, coding, writing, user interface design, organizing, and many others. 

To improve software that you and others use often
Most open source contributors almost always use software before they make a contribution to it. By using it, you are able to identify various bugs and flaws, making it easy for you to make contributions aimed at fixing the bugs. In so doing, you get the software to work better, making work easier for you and for numerous other users who are going to use it with and after you once it is fixed. 

In addition, you save others a lot of time and resources, especially those with limited skills and the ability to identify and fix bugs in software. 

To teach and learn from others